requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
streamlit
requests
openai
pandas
plotly

# Update these lines in requirements.txt
bcrypt
